Add: JWT authentication

+++ src/main/java/com/bankApp/dao/ClerkDaoImpl.java
@@ -1,6 +1,7 @@
 package com.bankApp.dao;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +9,6 @@ import com.bankApp.entities.Clerk;
 import com.bankApp.exceptions.BankEmployeeNotFoundException;
 
 import jakarta.persistence.EntityManager;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.PersistenceException;
 
@@ -21,7 +21,7 @@ public class ClerkDaoImpl implements ClerkDao {
 	public Clerk getById(Integer id) {
 		Clerk clerk = em.find(Clerk.class, id);
 		if(clerk == null) {
-			throw new EntityNotFoundException("Clerk with id = "+id+ " not found.");
+			throw new BankEmployeeNotFoundException("Clerk with id = "+id+ " not found.");
 		}
 		return clerk;
 	}
@@ -43,8 +43,10 @@ public class ClerkDaoImpl implements ClerkDao {
 	@Override
 	public void deleteClerk(Integer id) {
 		Clerk clerk = em.find(Clerk.class, id);
+		if(clerk == null) {
+			throw new BankEmployeeNotFoundException("Clerk with id="+id+" not found.");
+		}
 		em.remove(clerk);
-		
 	}
 
 	@Override
@@ -57,6 +59,7 @@ public class ClerkDaoImpl implements ClerkDao {
 //		} catch (IllegalArgumentException ex) {
 //			throw new BankEmployeeNotFoundException("Clerk not found.");
 //		}
+
 		Clerk existing = em.find(Clerk.class, clerk.getClerkId());
 		if(existing == null) {
 			throw new BankEmployeeNotFoundException("Clerk not found.");
@@ -64,6 +67,15 @@ public class ClerkDaoImpl implements ClerkDao {
 		if(clerk.getClerkName() != null) {
 			existing.setClerkName(clerk.getClerkName());
 		}		
+	}
+
+	@Override
+	public Optional<Clerk> findByClerkName(String username) {
+		Clerk clerk = em.find(Clerk.class, username);
+		if(clerk == null) {
+			throw new BankEmployeeNotFoundException("Clerk with name = "+username+ " not found.");
+		}
+		return Optional.of(clerk);
 	}
 
 }
